refactor: voice-input 컴포넌트 분리 및 녹음 시작 사운드 개선
- voice-input.tsx에서 하위 컴포넌트 분리
  - StatusError: 마이크 권한/장치 오류 표시
  - StatusMessage: 녹음 상태 메시지 표시
  - CountdownOverlay: 카운트다운 오버레이
  - PlaybackControls: 녹음 재생 컨트롤
- input-section.tsx에서 ErrorMessage 컴포넌트 분리
- useRecorder에 onStart 콜백 추가하여 녹음 시작 후 딩 사운드 재생
  - 기존: playDing() → startRecording() (동시 실행으로 사운드 끊김)
  - 변경: startRecording(onStart) → onStart에서 playDing() (녹음 시작 후 재생)
