+++ apps/web/src/app/daily/questions/[questionId]/_components/voice-input.tsx
@@ -3,18 +3,8 @@
 import * as React from "react";
 import Waveform from "@/components/waveform/waveform";
 import { Button } from "@/components/button/button";
-import {
-  CheckCircle2,
-  LoaderCircle,
-  RotateCcw,
-  MicOff,
-  ShieldAlert,
-  AlertCircle,
-  Play,
-  Pause,
-} from "lucide-react";
-import { Slider } from "@/components/slider/slider";
-import useRecorder, { RecorderStatus } from "../_hooks/use-recorder";
+import { CheckCircle2, LoaderCircle, RotateCcw } from "lucide-react";
+import useRecorder from "../_hooks/use-recorder";
 import RecordButton from "./record-button";
 import useAudio from "@/hooks/use-audio";
 import {
@@ -23,7 +13,10 @@ import {
   uploadToStorage,
 } from "../_lib/audio-upload-api";
 import { AUDIO_CONFIG } from "../_constants/audio-config-constant";
-import { formatTime } from "../_lib/format-time";
+import { StatusError } from "./status-error";
+import { StatusMessage, RecordingState } from "./status-message";
+import { CountdownOverlay } from "./countdown-overlay";
+import { PlaybackControls } from "./playback-controls";
 
 interface VoiceInputProps {
   maxDurationSeconds?: number;
@@ -33,8 +26,6 @@ interface VoiceInputProps {
   setIsSubmitting: (value: boolean) => void;
   disabled?: boolean;
 }
-
-type RecordingState = "idle" | "countdown" | "recording" | "recorded";
 
 function VoiceInput({
   maxDurationSeconds = 300,
@@ -75,7 +66,6 @@ function VoiceInput({
   React.useEffect(() => {
     if (countdown === null) return;
 
-    // disabled 또는 isSubmitting 상태면 카운트다운 취소
     if (disabled || isSubmitting) {
       setCountdown(null);
       return;
@@ -83,10 +73,12 @@ function VoiceInput({
 
     if (countdown === 0) {
       setCountdown(null);
-      if (dingReady) {
-        playDing();
-      }
-      startRecording();
+      // 녹음이 실제로 시작된 후 딩 사운드 재생
+      startRecording(() => {
+        if (dingReady) {
+          playDing();
+        }
+      });
       return;
     }
 
@@ -111,32 +103,6 @@ function VoiceInput({
       stopTickTock();
     }
   }, [isInWarningZone, tickTockReady, playTickTock, stopTickTock]);
-
-  const wasPlayingBeforeSeekRef = React.useRef(false);
-
-  const handleSeekStart = () => {
-    wasPlayingBeforeSeekRef.current = isPlaying;
-    if (isPlaying) {
-      pausePlayback();
-    }
-  };
-
-  const handleSeekEnd = (value: number[]) => {
-    seekTo(value[0]);
-    const shouldResume = wasPlayingBeforeSeekRef.current;
-    wasPlayingBeforeSeekRef.current = false;
-    if (shouldResume) {
-      playRecording();
-    }
-  };
-
-  const handleSeekCancel = () => {
-    const shouldResume = wasPlayingBeforeSeekRef.current;
-    wasPlayingBeforeSeekRef.current = false;
-    if (shouldResume) {
-      playRecording();
-    }
-  };
 
   const recordingState: RecordingState = isCountingDown
     ? "countdown"
@@ -207,44 +173,16 @@ function VoiceInput({
 
           <div className="max-w-sm w-full px-12">
             {hasRecorded ? (
-              <div className="flex items-center gap-2 md:gap-3 h-20 md:h-30">
-                <Button
-                  type="button"
-                  size="icon-lg"
-                  onClick={isPlaying ? pausePlayback : playRecording}
-                  disabled={isSubmitting}
-                  aria-label={isPlaying ? "일시정지" : "재생"}
-                  className="rounded-full shrink-0"
-                  aria-pressed={isPlaying}
-                >
-                  {isPlaying ? (
-                    <Pause className="w-5 h-5 fill-current" />
-                  ) : (
-                    <Play className="w-5 h-5 ml-0.5 fill-current" />
-                  )}
-                </Button>
-                <div className="flex-1 flex items-center gap-2 md:gap-3 min-w-0">
-                  <Slider
-                    value={[playbackTime]}
-                    min={0}
-                    max={duration || elapsedSeconds}
-                    step={0.1}
-                    onPointerDown={handleSeekStart}
-                    onPointerUp={handleSeekCancel}
-                    onPointerLeave={handleSeekCancel}
-                    onValueChange={([value]) => seekTo(value)}
-                    onValueCommit={handleSeekEnd}
-                    disabled={isSubmitting}
-                    className="flex-1"
-                    aria-label="녹음 재생 위치"
-                    aria-valuetext={`${formatTime(Math.floor(playbackTime))} / ${formatTime(Math.floor(duration || elapsedSeconds))}`}
-                  />
-                  <span className="text-xs md:text-sm tabular-nums text-muted-foreground shrink-0">
-                    {formatTime(Math.floor(playbackTime))} /{" "}
-                    {formatTime(Math.floor(duration || elapsedSeconds))}
-                  </span>
-                </div>
-              </div>
+              <PlaybackControls
+                isPlaying={isPlaying}
+                playbackTime={playbackTime}
+                duration={duration}
+                elapsedSeconds={elapsedSeconds}
+                isSubmitting={isSubmitting}
+                onPlay={playRecording}
+                onPause={pausePlayback}
+                onSeek={seekTo}
+              />
             ) : (
               <Waveform historyRef={historyRef} className="h-20 md:h-30" />
             )}
@@ -287,129 +225,6 @@ function VoiceInput({
           </div>
         </>
       )}
-    </div>
-  );
-}
-
-interface StatusErrorProps {
-  status: RecorderStatus;
-}
-
-function StatusError({ status }: StatusErrorProps) {
-  const config = {
-    unsupported: {
-      icon: AlertCircle,
-      title: "지원되지 않는 브라우저",
-      description:
-        "이 브라우저는 음성 녹음을 지원하지 않습니다. Chrome, Safari 등 최신 브라우저를 이용해주세요.",
-    },
-    permission_denied: {
-      icon: ShieldAlert,
-      title: "마이크 권한 필요",
-      description:
-        "음성 녹음을 위해 마이크 권한이 필요합니다. 브라우저 설정에서 마이크 권한을 허용해주세요.",
-    },
-    no_device: {
-      icon: MicOff,
-      title: "마이크를 찾을 수 없음",
-      description:
-        "연결된 마이크 장치가 없습니다. 마이크를 연결하고 다시 시도해주세요.",
-    },
-  } as const;
-
-  const current = config[status as keyof typeof config];
-
-  if (!current) return null;
-
-  const Icon = current.icon;
-
-  return (
-    <div className="flex flex-col items-center justify-center py-12 text-center px-6">
-      <div className="w-16 h-16 mb-4 rounded-full bg-red-50 flex items-center justify-center">
-        <Icon className="w-8 h-8 text-red-500" />
-      </div>
-      <h3 className="text-lg font-semibold text-gray-900 mb-2">
-        {current.title}
-      </h3>
-      <p className="text-sm text-gray-500 max-w-sm">{current.description}</p>
-    </div>
-  );
-}
-
-interface StatusMessageProps {
-  state: RecordingState;
-  elapsedSeconds: number;
-  maxDurationSeconds: number;
-}
-
-function StatusMessage({
-  state,
-  elapsedSeconds,
-  maxDurationSeconds,
-}: StatusMessageProps) {
-  const showIdle = state === "idle" || state === "countdown";
-
-  return (
-    <div className="text-center px-4 flex items-center justify-center">
-      {state === "recording" && (
-        <div
-          key="recording"
-          className="animate-in fade-in slide-in-from-bottom-2 duration-300 space-y-3"
-        >
-          <div className="pt-1">
-            <div className="text-4xl md:text-4xl font-semibold tabular-nums text-center tracking-tight flex items-end justify-center gap-1">
-              <div>{formatTime(elapsedSeconds)} </div>
-              <div className="text-base md:text-lg text-muted-foreground font-normal">
-                / {formatTime(maxDurationSeconds)}
-              </div>
-            </div>
-            <p className="text-muted-foreground text-center text-sm mt-1">
-              녹음 중...
-            </p>
-          </div>
-        </div>
-      )}
-
-      {showIdle && (
-        <div key="idle">
-          <h3 className="text-xl md:text-2xl font-bold mb-2 md:mb-3">
-            답변 시작
-          </h3>
-          <p className="text-muted-foreground text-sm md:text-base">
-            버튼을 눌러 녹음을 시작하세요.
-          </p>
-        </div>
-      )}
-
-      {state === "recorded" && (
-        <div
-          key="recorded"
-          className="animate-in fade-in slide-in-from-bottom-2 duration-300"
-        >
-          <h3 className="text-xl md:text-2xl font-bold mb-2 md:mb-3">
-            녹음 완료
-          </h3>
-          <p className="text-muted-foreground text-sm md:text-base">
-            답변을 제출하고 채점 결과를 확인해보세요
-          </p>
-        </div>
-      )}
-    </div>
-  );
-}
-
-function CountdownOverlay({ countdown }: { countdown: number | null }) {
-  if (countdown === null) return null;
-
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center animate-in fade-in duration-200">
-      <div className="absolute inset-0 bg-black/40 backdrop-blur-sm" />
-      <div
-        key={countdown}
-        className="relative text-8xl md:text-9xl font-bold text-teal-300 tabular-nums animate-countdown-spring"
-      >
-        {countdown}
-      </div>
     </div>
   );
 }
